Rename LogStyler to RoboLogStyle, and factor out CustomLogStyle.
